n1sdp: fix inconsistent logs in n1sdp modules
This patch fixes few 64-bit formatting issues in n1sdp_system
module and DIMM part number logging issue in n1sdp_dmc620
module.

+++ product/n1sdp/module/n1sdp_system/include/mod_n1sdp_system.h
@@ -41,7 +41,7 @@
 #define AP_SCP_SRAM_OFFSET            UINT32_C(0xA0000000)
 
 /*! AP Cores Reset Address in SCP Address Space */
-#define AP_CORE_RESET_ADDR            UINT64_C(0xA4040000)
+#define AP_CORE_RESET_ADDR            UINT32_C(0xA4040000)
 
 /*! Address translation enable bit */
 #define ADDR_TRANS_EN                 UINT32_C(0x1)
